+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Incoming API: View all incoming transactions, creating and deleting incoming transactions, updating package type of an existing incoming
 - Outgoing API: View all outgoing transactions, creating and deleting outgoing transactions, updating type of payment for an existing outgoing transaction
 - Budget API: View all budget transactions, creating and deleting budget, updating budget information
+- Organization API: View all, create, delete, update
 
 ### Changed
 - Path and functions in financial_transaction.php